requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
mistralai
aiofiles
